// clients/current/captive-portal/src/infrastructure/api/schema.d.ts
export interface components {
    schemas: {
// ...
        PortalSiteResponse: {
            ssid?: string;
            displayName?: null | string;
            logoUrl?: string;
            /** Format: int32 */
            sortOrder?: number | string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}